requirements.txt
python-telegram-bot==13.7
python-dotenv==0.19.0
Django==3.2.7
environs==9.3.3
jsonfield==3.1.0
django-phonenumber-field==5.2.0
phonenumbers==8.12.36
psycopg2==2.9.9
requests==2.32.3
